update component with simpler hooks
