+++ packages/template-retail-react-app/app/pages/checkout/partials/shipping-multi-address.jsx
@@ -46,90 +46,8 @@ import {useCheckout} from '@salesforce/retail-react-app/app/pages/checkout/util/
 import {usePickupShipment} from '@salesforce/retail-react-app/app/hooks/use-pickup-shipment'
 import {useAddressManagement} from '@salesforce/retail-react-app/app/hooks/use-address-management'
 import {useAddressForm} from '@salesforce/retail-react-app/app/hooks/use-address-form'
+import {useShipmentProcessing} from '@salesforce/retail-react-app/app/hooks/use-shipment-processing'
 import ProductAddressSelectionCard from '@salesforce/retail-react-app/app/components/product-address-selection-card'
-
-const MultiShippingItemAttributes = ({variant, includeQuantity = true}) => {
-    const {formatMessage} = useIntl()
-    const variationAttributes = variant?.variationAttributes || []
-    const variationValues = variant?.variationValues || {}
-    return (
-        <List
-            spacing={1.5}
-            flex={1}
-            aria-label={formatMessage({
-                id: 'shipping_multi_address.product_attributes.label',
-                defaultMessage: 'Product attributes'
-            })}
-        >
-            {variationAttributes &&
-                variationAttributes.length > 0 &&
-                variationAttributes.map((attr) => {
-                    const value = attr.values?.find((v) => v.value === variationValues[attr.id])
-                    return (
-                        <ListItem key={attr.id}>
-                            <Text lineHeight={1} color="gray.700" fontSize="sm">
-                                {attr.name || attr.id}: {value?.name || value?.value || ''}
-                            </Text>
-                        </ListItem>
-                    )
-                })}
-            {includeQuantity && (
-                <ListItem>
-                    <Text lineHeight={1} color="gray.700" fontSize="sm">
-                        {formatMessage({
-                            id: 'shipping_multi_address.quantity.label',
-                            defaultMessage: 'Quantity'
-                        })}
-                        : {variant.quantity}
-                    </Text>
-                </ListItem>
-            )}
-        </List>
-    )
-}
-
-MultiShippingItemAttributes.propTypes = {
-    variant: PropTypes.object.isRequired,
-    includeQuantity: PropTypes.bool
-}
-
-const AddressForm = ({item, form, onSubmit, onCancel}) => {
-    const saveButtonLabel = defineMessage({
-        defaultMessage: 'Save',
-        id: 'shipping_address_form.button.save'
-    })
-    return (
-        <Box position="relative" bg="white" padding={6} width="100%">
-            {form.formState.isSubmitting && <LoadingSpinner />}
-            <form
-                data-testid="address-form"
-                onSubmit={form.handleSubmit(async (data) => {
-                    await onSubmit(data, form, item.itemId)
-                })}
-            >
-                <Stack spacing={6} width="100%">
-                    {form.formState.errors?.global && (
-                        <Alert status="error">
-                            <AlertIcon color="red.600" boxSize={4} />
-                            <Text fontSize="sm" ml={3}>
-                                {form.formState.errors.global.message}
-                            </Text>
-                        </Alert>
-                    )}
-                    <AddressFields form={form} />
-                    <FormActionButtons onCancel={onCancel} saveButtonLabel={saveButtonLabel} />
-                </Stack>
-            </form>
-        </Box>
-    )
-}
-
-AddressForm.propTypes = {
-    item: PropTypes.object.isRequired,
-    form: PropTypes.object.isRequired,
-    onSubmit: PropTypes.func.isRequired,
-    onCancel: PropTypes.func.isRequired
-}
 
 const ShippingMultiAddress = ({
     basket,
@@ -142,24 +60,11 @@ const ShippingMultiAddress = ({
     const {currency} = useCurrency()
     const {STEPS, goToStep} = useCheckout()
     const showToast = useToast()
-    const {isCurrentShippingMethodPickup} = usePickupShipment(basket)
-    const {
-        findDeliveryShipmentWithSameAddress,
-        findUnusedDeliveryShipment,
-        createNewDeliveryShipmentWithAddress,
-        updateDeliveryAddressForShipment,
-        moveItemsToDeliveryShipment,
-        removeEmptyShipments,
-        areAddressesEqual
-    } = useMultiship(basket)
 
-    // Filter out pickup items - only show delivery items
-    const deliveryItems =
-        basket?.productItems?.filter((item) => {
-            const shipment = basket?.shipments?.find((s) => s.shipmentId === item.shipmentId)
-            return !isCurrentShippingMethodPickup(shipment?.shippingMethod)
-        }) || []
-    const productIds = deliveryItems.map((item) => item.productId).join(',')
+    const {areAddressesEqual} = useMultiship(basket)
+    const addressManagement = useAddressManagement(basket)
+    const productIds = addressManagement.deliveryItems.map((item) => item.productId).join(',')
+
     const {
         data: productsMap,
         isLoading: productsLoading,
@@ -185,14 +90,6 @@ const ShippingMultiAddress = ({
     } = useCurrentCustomer()
 
     const {
-        availableAddresses,
-        selectedAddresses,
-        setAddressesForItems,
-        addGuestAddress,
-        isGuest: isGuestUser
-    } = useAddressManagement(basket, deliveryItems)
-
-    const {
         form: addressForm,
         showForm: showAddAddressForm,
         isSubmitting: isFormSubmitting,
@@ -202,28 +99,28 @@ const ShippingMultiAddress = ({
         isAddressFormOpen,
         formErrors
     } = useAddressForm(
-        addGuestAddress,
-        isGuestUser,
-        setAddressesForItems,
-        availableAddresses,
-        deliveryItems,
+        addressManagement.addGuestAddress,
+        addressManagement.isGuest,
+        addressManagement.setAddressesForItems,
+        addressManagement.availableAddresses,
+        addressManagement.deliveryItems,
         areAddressesEqual
     )
 
+    const shipmentProcessing = useShipmentProcessing(basket)
+
     const [isSubmitting, setIsSubmitting] = useState(false)
 
-    // guest addresses for guests & customer addresses for registered users
-    const finalAddresses = availableAddresses
+    const addresses = addressManagement.availableAddresses
 
-    // Unified loading state - for guests, only check products loading since they may n't have addresses
-    const isLoading = (isGuestUser ? false : customerLoading) || productsLoading
+    // for guests, only check products loading since they may not have addresses yet
+    const isLoading = (addressManagement.isGuest ? false : customerLoading) || productsLoading
 
-    // Check if all product items have an address selected
-    const allShipmentsHaveAddress = (deliveryItems ?? []).every(
-        (item) => selectedAddresses[item.itemId]
+    const allShipmentsHaveAddress = addressManagement.deliveryItems.every(
+        (item) => addressManagement.selectedAddresses[item.itemId]
     )
 
-    if (!deliveryItems.length) {
+    if (!addressManagement.deliveryItems.length) {
         return (
             <Center
                 p={8}
@@ -290,65 +187,7 @@ const ShippingMultiAddress = ({
     const handleSubmit = async () => {
         setIsSubmitting(true)
         try {
-            // Based on the shopper's selected addresses, create a map of unique addressIds and their associated items
-            const addressToItemsMap = {}
-            let basketAfterItemMoves = null
-
-            deliveryItems.forEach((item) => {
-                // Defaults to the first address if no address is selected
-                const addressId = selectedAddresses[item.itemId] || finalAddresses[0]?.addressId
-                const address = finalAddresses.find((addr) => addr.addressId === addressId)
-
-                // If there is an existing shipment with the same address, use it in the next step
-                const shipmentIdWithSameAddress = findDeliveryShipmentWithSameAddress(
-                    basket,
-                    address
-                )
-
-                if (!addressToItemsMap[addressId]) {
-                    addressToItemsMap[addressId] = {
-                        address: address,
-                        items: [],
-                        shipmentId: shipmentIdWithSameAddress
-                    }
-                }
-                addressToItemsMap[addressId].items.push(item)
-            })
-
-            // For each unique address, if there is no usable existing shipment, create a new one.
-            for (const [addressId, data] of Object.entries(addressToItemsMap)) {
-                const {address, items, shipmentId: existingShipmentId} = data
-
-                let targetShipmentId = existingShipmentId
-                if (!targetShipmentId) {
-                    const targetShipment = findUnusedDeliveryShipment(
-                        basket,
-                        Object.values(addressToItemsMap).map((d) => d.shipmentId)
-                    )
-                    targetShipmentId = targetShipment?.shipmentId
-                    if (targetShipmentId) {
-                        await updateDeliveryAddressForShipment(targetShipmentId, address)
-                    } else {
-                        targetShipmentId = await createNewDeliveryShipmentWithAddress(
-                            basket,
-                            address
-                        )
-                    }
-                }
-                // Set the shipmentId for the unique address
-                addressToItemsMap[addressId].shipmentId = targetShipmentId
-                // Move items to the new shipment if needed.
-                const itemsToMove = items.filter((item) => item.shipmentId !== targetShipmentId)
-                if (itemsToMove.length > 0) {
-                    basketAfterItemMoves = await moveItemsToDeliveryShipment(
-                        itemsToMove,
-                        targetShipmentId
-                    )
-                }
-            }
-            // Remove any empty shipments.
-            await removeEmptyShipments(basketAfterItemMoves || basket)
-
+            await shipmentProcessing.processShipments(addressManagement.deliveryItems, addressManagement.selectedAddresses, addresses)
             goToStep(STEPS.SHIPPING_OPTIONS)
         } catch (error) {
             showToast({
@@ -375,7 +214,7 @@ const ShippingMultiAddress = ({
                     w="100%"
                 >
                     <VStack spacing={2} w="100%" h="100%">
-                        {deliveryItems.map((item) => {
+                        {addressManagement.deliveryItems.map((item) => {
                             const productDetail = productsMap?.[item.productId] || {}
                             const variant = {...item, ...productDetail}
                             const image = findImageGroupBy(productDetail.imageGroups, {
@@ -393,11 +232,11 @@ const ShippingMultiAddress = ({
                                     productDetail={productDetail}
                                     imageUrl={imageUrl}
                                     addressKey={addressKey}
-                                    selectedAddressId={selectedAddresses[addressKey]}
-                                    availableAddresses={finalAddresses}
-                                    isGuestUser={isGuestUser}
+                                    selectedAddressId={addressManagement.selectedAddresses[addressKey]}
+                                    availableAddresses={addresses}
+                                    isGuestUser={addressManagement.isGuest}
                                     customerLoading={customerLoading}
-                                    onAddressSelect={setAddressesForItems}
+                                    onAddressSelect={addressManagement.setAddressesForItems}
                                     onAddNewAddress={openForm}
                                     getPriceData={getPriceData}
                                     currency={currency}
